whatsapp-gemini-server/pyproject.toml
[project]
name = "whatsapp-gemini-server"
version = "0.1.0"
description = "MCP Server for WhatsApp integration with Gemini LLM"
dependencies = [
    "mcp>=1.2.0",
    "httpx>=0.24.0",
    "google-genai>=0.8.0",
    "twilio>=8.0.0",
    "python-dotenv>=1.0.0",
    "flask>=2.0.0",
]
requires-python = ">=3.10"
